+++ dart/test/scope_test.dart
@@ -268,7 +268,7 @@ void main() {
         ..setExtra('company-name', 'Dart Inc')
         ..setContexts('theme', 'material')
         ..addEventProcessor(
-          (event, hint) => event..tags.addAll({'page-locale': 'en-us'}),
+          (event, {hint}) => event..tags.addAll({'page-locale': 'en-us'}),
         );
 
       final updatedEvent = scope.applyToEvent(event, null);
@@ -416,8 +416,8 @@ class Fixture {
     return Scope(options);
   }
 
-  SentryEvent processor(SentryEvent event, dynamic hint) => null;
+  SentryEvent processor(SentryEvent event, {dynamic hint}) => null;
 
-  Breadcrumb beforeBreadcrumbCallback(Breadcrumb breadcrumb, dynamic hint) =>
+  Breadcrumb beforeBreadcrumbCallback(Breadcrumb breadcrumb, {dynamic hint}) =>
       null;
 }