+++ dart/test/sentry_client_test.dart
@@ -501,7 +501,7 @@ void main() {
     setUp(() {
       options = SentryOptions(dsn: fakeDsn);
       options.addEventProcessor(
-        (event, hint) => event
+        (event, {hint}) => event
           ..tags.addAll({'theme': 'material'})
           ..extra['host'] = '0.0.0.1'
           ..modules.addAll({'core': '1.0'})
@@ -537,10 +537,10 @@ void main() {
   });
 }
 
-SentryEvent beforeSendCallbackDropEvent(SentryEvent event, dynamic hint) =>
+SentryEvent beforeSendCallbackDropEvent(SentryEvent event, {dynamic hint}) =>
     null;
 
-SentryEvent beforeSendCallback(SentryEvent event, dynamic hint) {
+SentryEvent beforeSendCallback(SentryEvent event, {dynamic hint}) {
   return event
     ..tags.addAll({'theme': 'material'})
     ..extra['host'] = '0.0.0.1'