+++ flutter/test/load_contexts_integrations_test.dart
@@ -38,12 +38,12 @@ void main() {
     final options = SentryOptions()..dsn = fakeDsn;
     final hub = Hub(options);
 
-    loadContextsIntegration(options, _channel)(hub, options);
+    LoadContextsIntegration(_channel)(hub, options);
 
     expect(options.eventProcessors.length, 1);
 
     final e = SentryEvent();
-    final event = await options.eventProcessors.first(e, null);
+    final event = await options.eventProcessors.first(e);
 
     expect(called, true);
     expect(event.contexts.device.name, 'Device1');
@@ -67,7 +67,7 @@ void main() {
     final options = SentryOptions()..dsn = fakeDsn;
     final hub = Hub(options);
 
-    loadContextsIntegration(options, _channel)(hub, options);
+    LoadContextsIntegration(_channel)(hub, options);
 
     expect(options.eventProcessors.length, 1);
 
@@ -80,7 +80,7 @@ void main() {
         runtimes: [const SentryRuntime(name: 'eRT')])
       ..['theme'] = 'cuppertino';
     final e = SentryEvent(contexts: eventContexts);
-    final event = await options.eventProcessors.first(e, null);
+    final event = await options.eventProcessors.first(e);
 
     expect(called, true);
     expect(event.contexts.device.name, 'eDevice');
@@ -101,7 +101,7 @@ void main() {
       final options = SentryOptions()..dsn = fakeDsn;
       final hub = Hub(options);
 
-      loadContextsIntegration(options, _channel)(hub, options);
+      LoadContextsIntegration(_channel)(hub, options);
 
       final eventSdk = SdkVersion(
         name: 'sdk1',
@@ -110,7 +110,7 @@ void main() {
         packages: const [SentryPackage('event-package', '2.0')],
       );
       final e = SentryEvent(sdk: eventSdk);
-      final event = await options.eventProcessors.first(e, null);
+      final event = await options.eventProcessors.first(e);
 
       expect(
         event.sdk.packages.any((element) => element.name == 'native-package'),
@@ -132,10 +132,10 @@ void main() {
     final options = SentryOptions()..dsn = fakeDsn;
     final hub = Hub(options);
 
-    loadContextsIntegration(options, _channel)(hub, options);
+    LoadContextsIntegration(_channel)(hub, options);
 
     final e = SentryEvent();
-    final event = await options.eventProcessors.first(e, null);
+    final event = await options.eventProcessors.first(e);
 
     expect(event, isNotNull);
   });